training/onnx/requirements.txt
tensorflow-cpu==2.16.1
numpy==1.26.4
protobuf==3.20.3
onnx==1.17.0
tf2onnx==1.16.1
